Gateway Wrapper: enhance class hierarchy
Also add a metho to parse from protobuf message instead of payload
and generate generic_message instead of payload directly.
It is needed change for upcoming generic collection.
