+++ wrappers/python/wirepas_messaging/gateway/api/get_configs.py
@@ -18,7 +18,6 @@ from .config_helper import (
     set_config_ro,
     set_config_rw,
 )
-from .wirepas_exceptions import GatewayAPIParsingException
 
 
 class GetConfigsRequest(Request):
@@ -33,25 +32,18 @@ class GetConfigsRequest(Request):
         super(GetConfigsRequest, self).__init__(req_id=req_id, **kwargs)
 
     @classmethod
-    def from_payload(cls, payload):
-        message = wirepas_messaging.gateway.GenericMessage()
-        try:
-            message.ParseFromString(payload)
-        except Exception:
-            # Any Exception is promoted to Generic API exception
-            raise GatewayAPIParsingException("Cannot parse GetConfigsRequest payload")
-
+    def from_generic_message(cls, message):
         d = Request._parse_request_header(message.wirepas.get_configs_req.header)
         return cls(d["req_id"])
 
     @property
-    def payload(self):
+    def generic_message(self):
         message = wirepas_messaging.gateway.GenericMessage()
         # Fill the request header
         get_config = message.wirepas.get_configs_req
         get_config.header.CopyFrom(self._make_request_header())
 
-        return message.SerializeToString()
+        return message
 
 
 class GetConfigsResponse(Response):
@@ -70,14 +62,7 @@ class GetConfigsResponse(Response):
         self.configs = configs
 
     @classmethod
-    def from_payload(cls, payload):
-        message = wirepas_messaging.gateway.GenericMessage()
-        try:
-            message.ParseFromString(payload)
-        except Exception:
-            # Any Exception is promoted to Generic API exception
-            raise GatewayAPIParsingException("Cannot parse GetConfigsResponse payload")
-
+    def from_generic_message(cls, message):
         response = message.wirepas.get_configs_resp
 
         d = Response._parse_response_header(response.header)
@@ -97,7 +82,7 @@ class GetConfigsResponse(Response):
         return cls(d["req_id"], d["gw_id"], d["res"], configs)
 
     @property
-    def payload(self):
+    def generic_message(self):
         message = wirepas_messaging.gateway.GenericMessage()
 
         response = message.wirepas.get_configs_resp
@@ -110,4 +95,4 @@ class GetConfigsResponse(Response):
             set_config_rw(conf, config)
             set_config_ro(conf, config)
 
-        return message.SerializeToString()
+        return message
