+++ wrappers/python/wirepas_messaging/gateway/api/get_gw_info.py
@@ -12,8 +12,6 @@ import wirepas_messaging
 from .request import Request
 from .response import Response
 
-from .wirepas_exceptions import GatewayAPIParsingException
-
 
 class GetGatewayInfoRequest(Request):
     """
@@ -27,27 +25,18 @@ class GetGatewayInfoRequest(Request):
         super(GetGatewayInfoRequest, self).__init__(req_id=req_id, **kwargs)
 
     @classmethod
-    def from_payload(cls, payload):
-        message = wirepas_messaging.gateway.GenericMessage()
-        try:
-            message.ParseFromString(payload)
-        except Exception:
-            # Any Exception is promoted to Generic API exception
-            raise GatewayAPIParsingException(
-                "Cannot parse GetGatewayInfoRequest payload"
-            )
-
+    def from_generic_message(cls, message):
         d = Request._parse_request_header(message.wirepas.get_gateway_info_req.header)
         return cls(d["req_id"])
 
     @property
-    def payload(self):
+    def generic_message(self):
         message = wirepas_messaging.gateway.GenericMessage()
         # Fill the request header
         get_gateway_info = message.wirepas.get_gateway_info_req
         get_gateway_info.header.CopyFrom(self._make_request_header())
 
-        return message.SerializeToString()
+        return message
 
 
 class GetGatewayInfoResponse(Response):
@@ -81,16 +70,7 @@ class GetGatewayInfoResponse(Response):
         self.implemented_api_version = implemented_api_version
 
     @classmethod
-    def from_payload(cls, payload):
-        message = wirepas_messaging.gateway.GenericMessage()
-        try:
-            message.ParseFromString(payload)
-        except Exception:
-            # Any Exception is promoted to Generic API exception
-            raise GatewayAPIParsingException(
-                "Cannot parse GetGatewayInfoResponse payload"
-            )
-
+    def from_generic_message(cls, message):
         response = message.wirepas.get_gateway_info_resp
 
         d = Response._parse_response_header(response.header)
@@ -106,7 +86,7 @@ class GetGatewayInfoResponse(Response):
         )
 
     @property
-    def payload(self):
+    def generic_message(self):
         message = wirepas_messaging.gateway.GenericMessage()
 
         response = message.wirepas.get_gateway_info_resp
@@ -123,4 +103,4 @@ class GetGatewayInfoResponse(Response):
         if self.implemented_api_version is not None:
             response.info.implemented_api_version = self.implemented_api_version
 
-        return message.SerializeToString()
+        return message
