+++ wrappers/python/wirepas_messaging/gateway/api/request.py
@@ -10,8 +10,10 @@
 import random
 import wirepas_messaging
 
+from .generic_message import GenericMessage
 
-class Request(object):
+
+class Request(GenericMessage):
     """
     Request
 
@@ -27,14 +29,6 @@ class Request(object):
         if req_id is None:
             req_id = random.getrandbits(64)
         self.req_id = req_id
-
-    def __str__(self):
-        return str(self.__dict__)
-
-    @property
-    def payload(self):
-        """ Implement how to serialize child Event classes """
-        raise NotImplementedError()
 
     def _make_request_header(self):
         """ Creates the generic messaging header """
