+++ wrappers/python/wirepas_messaging/gateway/api/set_config.py
@@ -20,7 +20,6 @@ from .config_helper import (
     parse_config_ro,
     set_config_ro,
 )
-from .wirepas_exceptions import GatewayAPIParsingException
 
 
 class SetConfigRequest(Request):
@@ -54,14 +53,7 @@ class SetConfigRequest(Request):
         self.new_config = new_config
 
     @classmethod
-    def from_payload(cls, payload):
-        message = wirepas_messaging.gateway.GenericMessage()
-        try:
-            message.ParseFromString(payload)
-        except Exception:
-            # Any Exception is promoted to Generic API exception
-            raise GatewayAPIParsingException("Cannot parse SetConfigRequest payload")
-
+    def from_generic_message(cls, message):
         req = message.wirepas.set_config_req
 
         d = Request._parse_request_header(req.header)
@@ -77,7 +69,7 @@ class SetConfigRequest(Request):
         return cls(req.config.sink_id, new_config, d["req_id"])
 
     @property
-    def payload(self):
+    def generic_message(self):
         message = wirepas_messaging.gateway.GenericMessage()
         # Fill the request header
         set_config = message.wirepas.set_config_req
@@ -87,7 +79,7 @@ class SetConfigRequest(Request):
         set_config_rw(set_config.config, self.new_config)
         set_config_keys(set_config.config.keys, self.new_config)
 
-        return message.SerializeToString()
+        return message
 
 
 class SetConfigResponse(Response):
@@ -108,14 +100,7 @@ class SetConfigResponse(Response):
         self.config = config
 
     @classmethod
-    def from_payload(cls, payload):
-        message = wirepas_messaging.gateway.GenericMessage()
-        try:
-            message.ParseFromString(payload)
-        except Exception:
-            # Any Exception is promoted to Generic API exception
-            raise GatewayAPIParsingException("Cannot parse SetConfigResponse payload")
-
+    def from_generic_message(cls, message):
         response = message.wirepas.set_config_resp
 
         d = Response._parse_response_header(response.header)
@@ -131,7 +116,7 @@ class SetConfigResponse(Response):
         return cls(d["req_id"], d["gw_id"], d["res"], d["sink_id"], new_config)
 
     @property
-    def payload(self):
+    def generic_message(self):
         message = wirepas_messaging.gateway.GenericMessage()
 
         response = message.wirepas.set_config_resp
@@ -142,4 +127,4 @@ class SetConfigResponse(Response):
             set_config_rw(response.config, self.config)
             set_config_ro(response.config, self.config)
 
-        return message.SerializeToString()
+        return message
