+++ wrappers/python/wirepas_messaging/gateway/api/upload_scratchpad.py
@@ -11,7 +11,6 @@ import wirepas_messaging
 
 from .request import Request
 from .response import Response
-from .wirepas_exceptions import GatewayAPIParsingException
 
 
 class UploadScratchpadRequest(Request):
@@ -30,16 +29,7 @@ class UploadScratchpadRequest(Request):
         self.scratchpad = scratchpad
 
     @classmethod
-    def from_payload(cls, payload):
-        message = wirepas_messaging.gateway.GenericMessage()
-        try:
-            message.ParseFromString(payload)
-        except Exception:
-            # Any Exception is promoted to Generic API exception
-            raise GatewayAPIParsingException(
-                "Cannot parse UploadScratchpadRequest payload"
-            )
-
+    def from_generic_message(cls, message):
         req = message.wirepas.upload_scratchpad_req
 
         d = Request._parse_request_header(req.header)
@@ -52,7 +42,7 @@ class UploadScratchpadRequest(Request):
         return cls(req.seq, d["sink_id"], d["req_id"], scratchpad)
 
     @property
-    def payload(self):
+    def generic_message(self):
         message = wirepas_messaging.gateway.GenericMessage()
         # Fill the request header
         req = message.wirepas.upload_scratchpad_req
@@ -62,7 +52,7 @@ class UploadScratchpadRequest(Request):
         if self.scratchpad is not None:
             req.scratchpad = self.scratchpad
 
-        return message.SerializeToString()
+        return message
 
 
 class UploadScratchpadResponse(Response):
@@ -82,16 +72,7 @@ class UploadScratchpadResponse(Response):
         )
 
     @classmethod
-    def from_payload(cls, payload):
-        message = wirepas_messaging.gateway.GenericMessage()
-        try:
-            message.ParseFromString(payload)
-        except Exception:
-            # Any Exception is promoted to Generic API exception
-            raise GatewayAPIParsingException(
-                "Cannot parse UploadScratchpadResponse payload"
-            )
-
+    def from_generic_message(cls, message):
         response = message.wirepas.upload_scratchpad_resp
 
         d = Response._parse_response_header(response.header)
@@ -99,10 +80,10 @@ class UploadScratchpadResponse(Response):
         return cls(d["req_id"], d["gw_id"], d["res"], d["sink_id"])
 
     @property
-    def payload(self):
+    def generic_message(self):
         message = wirepas_messaging.gateway.GenericMessage()
 
         response = message.wirepas.upload_scratchpad_resp
         response.header.CopyFrom(self._make_response_header())
 
-        return message.SerializeToString()
+        return message
